requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.2
click==7.1.2
colorama==0.4.4
Flask==1.1.4
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.4
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
slack==0.0.2
slackclient==2.9.3
slackeventsapi==2.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==1.0.1
yarl==1.6.3
